Revert "Merge branch 'version-2' into version-1"
This reverts commit 201ad7d58cad5a5b31985c9a1d66ff4449307e81, reversing
changes made to 68c87b8f41314e4d573fd01a08cc73ffdb8a24f3.
