+++ src/App.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { Router } from "./routing";
 import'./global.scss'
