+++ src/Components/CreateTodo/CreateTodo.jsx
@@ -1,23 +1,25 @@
-import { useState, ChangeEvent, FormEvent } from "react";
+import React from 'react'
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { createTask } from '../../api.ts'
+import { createTask } from '../../../api';
 
 import './index.scss'
 
 const CreateTodo = () => {
+
     const navigate = useNavigate();
     const [title, setTitle] = useState("");
     const [description, setDescription] = useState("");
 
-    const handleTitleChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const handleTitleChange = (e) => {
         setTitle(e.target.value);
     };
 
-    const handleDescriptionChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
+    const handleDescriptionChange = (e) => {
         setDescription(e.target.value);
     };
 
-    const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
+    const handleSubmit = async (e) => {
         e.preventDefault();
 
         // Cria tarefa
@@ -30,6 +32,7 @@ const CreateTodo = () => {
         await createTask(newTask);
         navigate("/");
     };
+
 
     return (
         <div className="form">
@@ -56,11 +59,13 @@ const CreateTodo = () => {
                 </div>
 
                 <div className="form__BTsubmit">
-                    <button type="submit">Criar Tarefa</button>
+                    <button type="submit">Criar Tarefa</button>                
                 </div>
             </form>
         </div>
     );
 }
 
-export default CreateTodo;
+export default CreateTodo
+
+
