+++ src/Components/ViewTodo/TodoItem/TodoItem.jsx
@@ -1,21 +1,14 @@
-import { useState } from "react";
+import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { BsInfoCircleFill, BsCheck2Square } from "react-icons/bs";
 import { FaRegSquare } from "react-icons/fa";
 import { FaTrash } from "react-icons/fa";
 import { MdEdit } from "react-icons/md";
-import { updateTask } from "../../../api"; 
+import { updateTask } from "../../../../api"; 
 
 import './index.scss';
 
-export interface TaskType {
-    id: string;
-    title: string;
-    description: string;
-    completed: boolean;
-}
-
-const TodoItem = ({ id, title, description, completed }: TaskType ) => {
+const TodoItem = ({ id, title, description, completed }) => {
     const navigate = useNavigate();
     const [isChecked, setIsChecked] = useState(completed);
