+++ src/Pages/DetailTodo/TodoDetail/TodoDetail.jsx
@@ -1,15 +1,7 @@
-import { ReactNode } from 'react'
+import React from 'react'
 import './index.scss';
 
-
-export interface TaskDetailType {
-    label: string;
-    value: ReactNode;
-    type: string;
-    onChange?: () => void;
-}
-
-const TodoDetail = ({ label, value, type, } : TaskDetailType) => {
+const TodoDetail = ({ label, value, type, onChange }) => {
     const id = `input${label}`;
 
     return (
@@ -18,7 +10,7 @@ const TodoDetail = ({ label, value, type, } : TaskDetailType) => {
             {type === "custom" ? (
                 <div className="custom-icon-container">{value}</div>
             ) : (
-                <input type={type} id={id} value={value as string}  readOnly />
+                <input type={type} id={id} value={value} readOnly />
             )}
         </div>
     );
