+++ src/Pages/VDeleteTodo/VDeleteTodo.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import DeleteTodo from '../../Components/DeleteTodo/DeleteTodo'
 import './index.scss'
