+++ src/routing/HomePage/HomePage.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import Header from './Header/Header'
 import { Outlet } from 'react-router-dom'
